cleanup: Crown of stars mote creation loop
The function was made when I just wanted to get something working, so
the names of functions were nonsensical and the loop for creating motes.

Updated the loop to use more descriptive variable names and improve readability.
Effectively condensing the loop to just use a "add mote" function call repeatedly.

Tested: I'm using the same creation code in my own module, and it works as expected.
Signed-off-by: aljames-arctic <bakanabaka.vtt@gmail.com>